Updates with requested changes.
Requested changes have been made, and the script has been made more robust as I noticed that when collecting the data during market hours, the time it takes to request the data is enough to create differences in the Price, Change, and Volume columns. Solution was to drop them from four of five dataframe.